Issue #1348: Added widgets gallery
