--- src/app/modules/dashboard/components/widget-menu/widget-menu.component.html
@@ -1,37 +1,0 @@
-<ul *transloco="let t; scope: 'dashboard/widget-menu'" nz-menu nzMode="vertical">
-  <ng-container *ngrxLet="menuDisplay$ as menu">
-
-    <ng-container *ngFor="let widget of menu.allCategory.widgets">
-      <li
-        (click)="selected.emit(widget.typeId)"
-        nz-menu-item
-      >
-        <span><i [nzType]="widget.icon" nz-icon nzTheme="outline"></i>&nbsp;{{widget.name}}</span>
-      </li>
-    </ng-container>
-
-    <ng-container *ngFor="let group of menu.groups">
-      <li [nzTitle]="t('dashboardWidgetMenu.categories.' + group.category, {fallback: group.category})" nz-submenu>
-        <ul>
-          <ng-container *ngFor="let widget of group.widgets">
-            <li
-              (click)="selected.emit(widget.typeId)"
-              nz-menu-item
-            >
-              <span><i [nzType]="widget.icon" nz-icon nzTheme="outline"></i>&nbsp;{{widget.name}}</span>
-            </li>
-          </ng-container>
-        </ul>
-      </li>
-    </ng-container>
-  </ng-container>
-
-  <ng-container *ngIf="showResetItem">
-    <li nz-menu-divider></li>
-    <li (click)="resetDashboard.emit()" nz-menu-item>
-      <span>
-        <i nz-icon nzTheme="outline" nzType="appstore"></i>{{t('dashboardWidgetMenu.standardView')}}
-      </span>
-    </li>
-  </ng-container>
-</ul>
