--- src/app/modules/dashboard/components/widget-menu/widget-menu.component.less
@@ -1,14 +1,0 @@
-li {
-  padding-left: 14px;
-  padding-right: 25px;
-
-  i {
-    margin-right: 7px;
-  }
-
-  ::ng-deep {
-    .ant-dropdown-menu-submenu-title {
-      padding-left: 0;
-    }
-  }
-}
