--- src/app/modules/dashboard/components/widget-menu/widget-menu.component.spec.ts
@@ -1,47 +1,0 @@
-import {ComponentFixture, TestBed} from '@angular/core/testing';
-
-import {WidgetMenuComponent} from './widget-menu.component';
-import {of} from "rxjs";
-import {
-  getTranslocoModule,
-  ngZorroMockComponents
-} from "../../../../shared/utils/testing";
-import { WidgetsMetaService } from "../../../../shared/services/widgets-meta.service";
-import { LetDirective } from "@ngrx/component";
-
-describe('WidgetMenuComponent', () => {
-  let component: WidgetMenuComponent;
-  let fixture: ComponentFixture<WidgetMenuComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      declarations: [
-        WidgetMenuComponent,
-        ...ngZorroMockComponents
-      ],
-      imports: [
-        getTranslocoModule(),
-        LetDirective
-      ],
-      providers: [
-        {
-          provide: WidgetsMetaService,
-          useValue: {
-            getWidgetsMeta: jasmine.createSpy('getWidgetsMeta').and.returnValue(of([]))
-          }
-        }
-      ]
-    })
-      .compileComponents();
-  });
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(WidgetMenuComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
