+++ src/app/modules/dashboard/components/widgets-gallery/widgets-gallery.component.less
@@ -1,0 +1,60 @@
+@import "theme-mixin";
+@import 'scrollable-mixin';
+
+.themeMixin({
+  .widget-gallery {
+    .gallery-item {
+      &:hover {
+        background-color: @select-item-selected-bg;
+      }
+    }
+
+    .divider {
+      color: @heading-color;
+      font-size: @font-size-lg;
+
+      &::after {
+        border-top: @border-width-base solid @divider-color;
+      }
+    }
+
+    .ant-drawer-body {
+      .scrollable(@scrollbar-color);
+    }
+  }
+});
+
+.widget-gallery {
+  .gallery-item {
+    min-height: 28px;
+  }
+
+  .divider {
+    display: flex;
+    margin: 14px 0;
+    font-weight: 500;
+    white-space: nowrap;
+    text-align: center;
+    border-top: 0;
+
+    .divider-text {
+      padding-right: 10px;
+    }
+
+    &::after {
+      position: relative;
+      top: 50%;
+      width: 100%;
+      border-top-color: inherit;
+      border-bottom: 0;
+      transform: translateY(50%);
+      content: '';
+    }
+  }
+
+  .ant-drawer-title {
+    -webkit-user-select: none;
+    -moz-user-select: none;
+    user-select: none;
+  }
+}
