+++ src/assets/i18n/dashboard/widgets-gallery/en.json
@@ -1,0 +1,12 @@
+﻿{
+  "title": "Widgets",
+  "standardView": "Standard View",
+
+  "categories": {
+    "all": "All",
+    "chartsAndOrderbooks": "Charts and Orderbooks",
+    "positionsTradesOrders": "Positions Trades and Orders",
+    "info": "Information Resources",
+    "details": "Detailed"
+  }
+}
