+++ src/styles.less
@@ -9,6 +9,9 @@
 @import 'color-utils';
 @import 'overflow-utils';
 @import 'z-index-utils';
+@import 'cursor-utils';
+@import 'font-utils';
+@import 'interactions-utils';
 
 html {
   height: 100%;
