+++ src/styles/themes/base.less
@@ -1,7 +1,5 @@
 ﻿@line-height-base: 1.66667;
 @mode: compact;
-@font-size-base: 12px;
-@font-size-lg: @font-size-base + 2px;
 
 // default paddings
 @default-padding-lg: 24px; // containers
