+++ src/styles/themes/dark.less
@@ -2,6 +2,9 @@
 @import './base';
 @import './dark-shared-colors';
 
+@font-size-base: 12px;
+@font-size-lg: @font-size-base + 2px;
+
 // header layout
 @layout-header-color: #E2E5E9;
 @layout-header-background: #141922;
