+++ src/styles/themes/default.less
@@ -2,6 +2,9 @@
 @import './base';
 @import './default-shared-colors';
 
+@font-size-base: 12px;
+@font-size-lg: @font-size-base + 2px;
+
 // header layout
 @layout-header-color: #FFFFFF;
 @layout-header-background: #677087;
