+++ src/styles/utils/cursor-utils.less
@@ -1,0 +1,5 @@
+﻿.cursor {
+  &-pointer {
+    cursor: pointer;
+  }
+}
