+++ src/styles/utils/interactions-utils.less
@@ -1,0 +1,19 @@
+﻿.user-select {
+  &-all {
+    -webkit-user-select: all;
+    -moz-user-select: all;
+    user-select: all;
+  }
+
+  &-auto {
+    -webkit-user-select: auto;
+    -moz-user-select: auto;
+    user-select: auto;
+  }
+
+  &-none {
+    -webkit-user-select: none;
+    -moz-user-select: none;
+    user-select: none;
+  }
+}
