+++ src/styles/utils/spacing-utils.less
@@ -1,4 +1,4 @@
-each(range(0, 10, 1), {
+each(range(0, 25, 1), {
   .m-@{value} {
     margin: @value * 1px;
   }
